Stepdefinitions and Steps class
The classes are meant to tie the business logic with code. Using gerkin syntac you can describe a feature with Given When Then.

In the Step Definition class we can pre set what we need to do , set a config file to create a user or the right set of conditions to manipulate data so on.

Steps class  we call endpoints and it's where we assert endpoints that we called in order to verify the behavior flow. It also makes it easier to make tests generic and reusable using methods from Steps in StepDef