Implementing core logic
CucumberTestSuite class is for debugging purposes

Endpoint enum where we hold all the endpoints variables

MobiquityClient is where the endpoints are called with SerenityRest. This part of the code would be relocated to a separate project in order to be used by Gatling and the Api test automated framework so that we only would write the code once and re-use it. Also its a single source of truth as long as you send the correct parameters the endpoint does what it exactly says e.g {createUser, deleteUser etc}

Environment Controller is to set environment - (in this case we only have a test environment so unnecessary methods weren't created because they aren't being used). It also handles setting up pre requisites to json objects that need to be passed in tests

Pojo folders generated from json structure  to be deserialized from json to java objects { User, Geo, Company, Comments, Address}